+++ app/validators/concerns/decidim/decidim_awesome/etiquette_validator_override.rb
@@ -14,9 +14,14 @@ module Decidim
 
         def validate_caps(record, attribute, value)
           awesome_config = awesome_config(record, "validate_#{attribute_without_locale(attribute)}_max_caps_percent")
-          return original_validate_caps(record, attribute, value) if awesome_config.nil?
 
-          percent = awesome_config.to_f
+          # original method does not take into account accents or anything else than A-Z
+          # return original_validate_caps_first(record, attribute, value) if awesome_config.nil?
+          percent = if awesome_config.nil?
+                      50
+                    else
+                      awesome_config.to_f
+                    end
           return if value.scan(/[[:upper:]]/).length < value.length * percent / 100
 
           record.errors.add(attribute, options[:message] || I18n.t("too_much_caps", scope: "decidim.decidim_awesome.validators", percent: percent.round))
@@ -24,9 +29,12 @@ module Decidim
 
         def validate_marks(record, attribute, value)
           awesome_config = awesome_config(record, "validate_#{attribute_without_locale(attribute)}_max_marks_together")
-          return original_validate_marks(record, attribute, value) if awesome_config.nil?
 
-          marks = awesome_config.to_i + 1
+          marks = if awesome_config.nil?
+                    2
+                  else
+                    awesome_config.to_i + 1
+                  end
           return if value.scan(/[!?¡¿]{#{marks},}/).empty?
 
           record.errors.add(attribute, options[:message] || :too_many_marks)
@@ -34,18 +42,21 @@ module Decidim
 
         def validate_caps_first(record, attribute, value)
           awesome_config = awesome_config(record, "validate_#{attribute_without_locale(attribute)}_start_with_caps")
-          return original_validate_caps_first(record, attribute, value) if awesome_config.nil?
-          return unless awesome_config
+
+          # original method does not take into account accents or anything else than A-Z
+          # return original_validate_caps_first(record, attribute, value) if awesome_config.nil?
+          return unless awesome_config || awesome_config.nil?
           return if value.scan(/\A[[:lower:]]{1}/).empty?
 
           record.errors.add(attribute, options[:message] || :must_start_with_caps)
         end
 
-        def awesome_config(record, var)
-          config = record.try(:awesome_config)&.config
-          return unless config.is_a?(Hash)
+        def record_awesome_config(record)
+          record.try(:awesome_config)
+        end
 
-          config[var.to_sym]
+        def awesome_config(record, var)
+          record_awesome_config(record)&.config&.[](var.to_sym)
         end
 
         def attribute_without_locale(attribute)